refactor(examples): give the nemotron weight layout a home of its own
`runtime_model.load` imported `mega_kernel` unconditionally, so every
implementation needed TileLang installed -- including the two that exist to not
need it. The layout it was reaching for (`PACK`, `Packed`, `pack_into`) is a
property of the model rather than of any backend, and it is pure torch.

It moves to `packing.py`, which imports no backend. `gen_kernel.py` emits an
import of it in place of the block it used to write out, so `mega_kernel.py` and
the handwritten CUDA path read one definition and cannot drift; a drifted layout
is a silently wrong weight, not an error. The generated kernel body is otherwise
unchanged (1564 -> 1505 lines, the difference being exactly that block).

`load` now builds a `mega_kernel.Runner` only for `NEMO_IMPL=mega`.

Verified: `compare_impls.py --a ops --b cuda` runs in a venv with no TileLang at
all and lands at 3.489e-2 against the 5.018e-2 bound `check_all.py` derives from
its own bf16 landing count, with the same argmax at every step.
